# cmd/data-aggregation/docs/swagger.yaml
basePath: /
definitions:
  httputil.Response:
    properties:
      error:
        type: string
    type: object
  models.Subscription:
    properties:
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  requests.CreateSubscriptionRequest:
    properties:
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  requests.SumSubscriptionRequest:
    properties:
      end_date:
        type: string
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  requests.UpdateSubscriptionRequest:
    properties:
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  responses.CreateSubscriptionResponse:
    properties:
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  responses.DeleteSubscriptionResponse:
    properties:
      id:
        type: integer
      message:
        type: string
    type: object
  responses.SumSubscriptionResponse:
    properties:
      total_sum:
        type: integer
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: show list of subscriptions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Subscription'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Show subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: create subscription
      parameters:
      - description: Subscription Info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.CreateSubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: delete subscription by id
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DeleteSubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: get subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Show subscription
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: update subscription by id
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: subscription info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CreateSubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/sum:
    post:
      consumes:
      - application/json
      description: sum price of all subscriptions for the selected period filtered
        by user_id and service_name
      parameters:
      - description: Subscription Info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.SumSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SumSubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Sum subscriptions
      tags:
      - subscriptions
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"